refactoring api_report.py to report.py
