+++ ross/__init__.py
@@ -4,7 +4,7 @@ import plotly.io as _pio
 
 import ross.plotly_theme
 
-from .api_report import *
+from .report import *
 from .bearing_seal_element import *
 from .disk_element import *
 from .materials import *
